refactor: factor out questions and use strong-typing for question args
